backend/requirements.txt
asgiref==3.8.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.0
Django==5.1.5
django-allauth==65.3.1
django-cors-headers==4.6.0
django-extensions==3.2.3
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
idna==3.10
iniconfig==2.0.0
packaging==24.2
pluggy==1.5.0
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
pytest==8.3.5
pytest-django==4.10.0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
sqlparse==0.5.3
urllib3==2.3.0
openai==0.28
spacy == 3.8.4
sentence-transformers == 3.4.1
scikit-learn == 1.6.1
django-storages == 1.14.5
boto3 == 1.37.7